refactor: fix issues
